Fix ErrorBoundary and module errors
The following errors were fixed:
- Errors related to `ErrorBoundary` and `RootErrorBoundary` not having `state` or `props`.
- Missing module errors for `pdf-parse` and `next-themes/dist/types`.
- Type errors in `accessible-heading`, `sheet`, and `sidebar` components.
- Type assignment errors in `Tips.tsx` and `PitchAnalysisService.ts`.

+++ src/services/PitchAnalysisService.ts
@@ -45,7 +45,7 @@ export class PitchAnalysisService {
       return;
     }
 
-    this.analyser.getFloatTimeDomainData(this.dataArray);
+    this.analyser.getFloatTimeDomainData(this.dataArray as Float32Array);
     
     // Process audio data for pitch detection
     if (this.onAnalysisCallback) {
